+++ Sources/RichTextKit/RichTextViewComponent+Alignment.swift
@@ -1,5 +1,5 @@
 //
-//  RichTextViewRepresentable+Alignment.swift
+//  RichTextViewComponent+Alignment.swift
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2022-05-29.
@@ -16,7 +16,7 @@ import UIKit
 import AppKit
 #endif
 
-public extension RichTextViewRepresentable {
+public extension RichTextViewComponent {
 
     /**
      Use the selected range (if any) or text position to get
@@ -45,7 +45,7 @@ public extension RichTextViewRepresentable {
     }
 }
 
-private extension RichTextViewRepresentable {
+private extension RichTextViewComponent {
 
     /**
      Set the text alignment at the current position.