+++ Sources/RichTextKit/_Deprecated/RichTextViewRepresentable.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "RichTextViewComponent")
+public typealias RichTextViewRepresentable = RichTextViewComponent